+++ src/service/provider/DataProvider.js
@@ -1,4 +1,5 @@
 import React, { useState, useEffect, createContext } from "react";
+import { getCubes } from "../data/apiCallers";
 
 const DataContext = createContext();
 
@@ -6,19 +7,7 @@ const DataProvider = ({ children }) => {
   const [data, setData] = useState([]);
 
   useEffect(() => {
-    const fetchData = async () => {
-      try {
-        const res = await fetch(" https://arkansas-api.datausa.io/cubes");
-        if (!res.ok) throw new Error("Failed getting data");
-
-        const data = await res.json();
-        setData(data.cubes);
-      } catch (error) {
-        console.error("Error fetching data:", error);
-      }
-    };
-
-    fetchData();
+    getCubes().then((data) => setData(data));
   }, []);
 
   return <DataContext.Provider value={data}>{children}</DataContext.Provider>;
